requirements.txt
pandas==1.4.2
azure-storage-blob==12.13.1
smart_open==5.1.0
numpy==1.21.5
nltk==3.7
scikit-learn==1.0.2
openpyxl==3.0.9
pickle==4.0